+++ test/JsonApiDotNetCoreExampleTests/IntegrationTests/ResourceInheritance/Models/Female.cs
@@ -1,0 +1,10 @@
+using JsonApiDotNetCore.Resources.Annotations;
+
+namespace JsonApiDotNetCoreExampleTests.IntegrationTests.ResourceInheritance
+{
+    public sealed class Female : Person
+    {
+        [Attr]
+        public string FemaleProperty { get; set; }
+    }
+}
