+++ test/JsonApiDotNetCoreExampleTests/IntegrationTests/ResourceInheritance/Models/Male.cs
@@ -1,0 +1,10 @@
+using JsonApiDotNetCore.Resources.Annotations;
+
+namespace JsonApiDotNetCoreExampleTests.IntegrationTests.ResourceInheritance
+{
+    public sealed class Male : Person
+    {
+        [Attr]
+        public string MaleProperty { get; set; }
+    }
+}
