+++ test/JsonApiDotNetCoreExampleTests/IntegrationTests/ResourceInheritance/Models/Person.cs
@@ -1,0 +1,6 @@
+using JsonApiDotNetCore.Resources;
+
+namespace JsonApiDotNetCoreExampleTests.IntegrationTests.ResourceInheritance
+{
+    public abstract class Person : Identifiable { }
+}
