+++ test/JsonApiDotNetCoreExampleTests/IntegrationTests/ResourceInheritance/Models/PlaceholderPerson.cs
@@ -1,0 +1,13 @@
+namespace JsonApiDotNetCoreExampleTests.IntegrationTests.ResourceInheritance
+{
+    public sealed class PlaceholderPerson
+    {
+        public int PlaceHolderId { get; set; }
+        
+        public Placeholder PlaceHolder { get; set; }
+
+        public int PersonId { get; set; }
+        
+        public Person Person { get; set; }
+    }
+}
